ethercatmcIndexerAxis.cpp: Support VelAct
Calculate the "actual velocity" and save it into the "-VelAct" PV.

If available, use the UTC time from the MCU:
Add ctrlLocal.oldSystemUTCtimeNsec and ctrlLocal.newSystemUTCtimeNsec
in the poller in ethercatmcIndexer.cpp

If available: remove the deleay  caused by the TCP/IP communication.
Use SystemUTCtimeNsec instead of the EPICS time in the IOC to
calculate the velocity.

Note: The motorRecord has a actual velocity in steps/second,
      this one is not supported

Note2: There is a filter when calculating the actual velocity,
       depending on the old and very old positions.
       This seems to work on our test crates.
       In the long run a real deadband handling is needed.

+++ ethercatmcApp/src/ethercatmcController.cpp
@@ -44,7 +44,7 @@ double ethercatmcgetNowTimeSecs(void)
 {
     epicsTimeStamp nowTime;
     epicsTimeGetCurrent(&nowTime);
-    return nowTime.secPastEpoch + (nowTime.nsec * 0.000000001);
+    return (double)nowTime.secPastEpoch + (nowTime.nsec * 0.000000001);
 }
 
 extern "C" const char *errStringFromErrId(int nErrorId)
